+++ src/py3createtorrent.py
@@ -2,7 +2,7 @@
 """
 Create torrents via command line!
 
-Copyright (C) 2010-2020 Robert Nitsch
+Copyright (C) 2010-2021 Robert Nitsch
 Licensed according to GPL v3.
 """
 
@@ -10,9 +10,11 @@ import argparse
 import concurrent.futures
 import datetime
 import hashlib
+import json
 import math
 import multiprocessing
 import os
+import pprint
 import re
 import sys
 import time
@@ -38,29 +40,11 @@ https://py3createtorrent.readthedocs.io/en/latest/""")
     print()
     raise
 
-__all__ = ['calculate_piece_length', 'get_files_in_directory', 'sha1_20', 'split_path']
-
-# #############
-# CONFIGURATION
-
-# Configure your tracker abbreviations here.
-TRACKER_ABBR = {
-    'opentrackr': 'udp://tracker.opentrackr.org:1337/announce',
-    'coppersurfer': 'udp://tracker.coppersurfer.tk:6969/announce',
-    'cyberia': 'udp://tracker.cyberia.is:6969/announce'
-}
-
-# Whether or not py3createtorrent is allowed to advertise itself through the torrents' comment fields.
-ADVERTISE = True
-
-BEST_TRACKERS_URL = "https://raw.githubusercontent.com/ngosang/trackerslist/master/trackers_best.txt"
-
-# /CONFIGURATION
-# ##############
+__all__ = ['calculate_piece_length', 'get_files_in_directory', 'sha1', 'split_path']
 
 # Do not touch anything below this line unless you know what you're doing!
 
-__version__ = '1.0.0.dev4'
+__version__ = '1.0.0b2'
 
 # Note:
 #  Kilobyte = kB  = 1000 Bytes
@@ -71,17 +55,76 @@ MIB = KIB * KIB
 VERBOSE = False
 
 
+class Config(object):
+    class InvalidConfigError(Exception):
+        pass
+
+    def __init__(self) -> None:
+        self.path = None  # type: Optional[str]
+        self.tracker_abbreviations = {
+            'opentrackr': 'udp://tracker.opentrackr.org:1337/announce',
+            'coppersurfer': 'udp://tracker.coppersurfer.tk:6969/announce',
+            'cyberia': 'udp://tracker.cyberia.is:6969/announce'
+        }
+        self.advertise = True  # type: Optional[bool]
+        self.best_trackers_url = "https://raw.githubusercontent.com/ngosang/trackerslist/master/trackers_best.txt"  # type: str
+
+    def get_path_to_config_file(self) -> str:
+        if self.path is None:
+            return os.path.join(os.path.expanduser('~'), '.py3createtorrent.cfg')
+        else:
+            return self.path
+
+    def load_config(self) -> None:
+        """
+        @throws json.JSONDecodeError if config file cannot be parsed as JSON
+        """
+        path = self.get_path_to_config_file()
+        printv("Path to config file: ", path)
+
+        if not os.path.isfile(path):
+            printv('Config file does not exist')
+            return
+
+        with open(path, 'r') as fh:
+            data = json.load(fh)
+
+        self.tracker_abbreviations = data.get('tracker_abbreviations', self.tracker_abbreviations)
+        self.advertise = data.get('advertise', self.advertise)
+        self.best_trackers_url = data.get('best_trackers_url', self.best_trackers_url)
+
+        # Validate the configuration.
+        for abbr, replacement in self.tracker_abbreviations.items():
+            if not isinstance(abbr, str):
+                raise Config.InvalidConfigError("Configuration error: invalid tracker abbreviation: '%s' "
+                                                "(must be a string instead)" % abbr)
+            if not isinstance(replacement, (str, list)):
+                raise Config.InvalidConfigError("Configuration error: invalid tracker abbreviation: '%s' "
+                                                "(must be a string or list of strings instead)" % str(replacement))
+
+        if not isinstance(self.best_trackers_url, str):
+            raise Config.InvalidConfigError("Configuration error: invalid best trackers url: %s "
+                                            "(must be a string)" % self.best_trackers_url)
+        if not self.best_trackers_url.startswith("http"):
+            raise Config.InvalidConfigError("Configuration error: invalid best trackers url: %s "
+                                            "(must be a http/https URL)" % self.best_trackers_url)
+
+        if not isinstance(self.advertise, bool):
+            raise Config.InvalidConfigError("Configuration error: invalid value for advertise: %s "
+                                            "(must be true/false)" % self.best_trackers_url)
+
+
 def printv(*args: Any, **kwargs: Any) -> None:
     """If VERBOSE is True, act as an alias for print. Else do nothing."""
     if VERBOSE:
         print(*args, **kwargs)
 
 
-def sha1_20(data: bytes) -> bytes:
-    """Return the first 20 bytes of the given data's SHA-1 hash."""
+def sha1(data: bytes) -> bytes:
+    """Return the given data's SHA-1 hash (= always 20 bytes)."""
     m = hashlib.sha1()
     m.update(data)
-    return m.digest()[:20]
+    return m.digest()
 
 
 def create_single_file_info(file: str, piece_length: int, include_md5: bool = True) -> Dict:
@@ -99,6 +142,7 @@ def create_single_file_info(file: str, piece_length: int, include_md5: bool = Tr
 
     # Total byte count.
     length = os.path.getsize(file)
+    assert length > 0, "empty file"
 
     # Concatenated 20byte sha1-hashes of all the file's pieces.
     piece_count = int(math.ceil(length / piece_length))
@@ -112,9 +156,9 @@ def create_single_file_info(file: str, piece_length: int, include_md5: bool = Tr
     def calculate_sha1_hash_for_piece(i, piece_data):
         count = len(piece_data)
         if count == piece_length:
-            pieces[i * 20:(i + 1) * 20] = sha1_20(piece_data)
+            pieces[i * 20:(i + 1) * 20] = sha1(piece_data)
         elif count != 0:
-            pieces[i * 20:(i + 1) * 20] = sha1_20(piece_data[:count])
+            pieces[i * 20:(i + 1) * 20] = sha1(piece_data[:count])
 
     MAX_FUTURES = max(2, multiprocessing.cpu_count() - 1)
     with concurrent.futures.ThreadPoolExecutor(max_workers=MAX_FUTURES) as executor:
@@ -136,8 +180,6 @@ def create_single_file_info(file: str, piece_length: int, include_md5: bool = Tr
             concurrent.futures.wait(futures)
 
     printv("done")
-
-    assert length > 0, "empty file"
 
     info = {
         'pieces': bytes(pieces),
@@ -185,9 +227,9 @@ def create_multi_file_info(directory: str, files: List[str], piece_length: int, 
     def calculate_sha1_hash_for_piece(i, piece_data):
         count = len(piece_data)
         if count == piece_length:
-            pieces[i * 20:(i + 1) * 20] = sha1_20(piece_data)
+            pieces[i * 20:(i + 1) * 20] = sha1(piece_data)
         elif count != 0:
-            pieces[i * 20:(i + 1) * 20] = sha1_20(piece_data[:count])
+            pieces[i * 20:(i + 1) * 20] = sha1(piece_data[:count])
 
     MAX_FUTURES = max(2, multiprocessing.cpu_count() - 1)
     with concurrent.futures.ThreadPoolExecutor(max_workers=MAX_FUTURES) as executor:
@@ -246,7 +288,7 @@ def create_multi_file_info(directory: str, files: List[str], piece_length: int, 
 
     # Don't forget to hash the last piece. (Probably the piece that has not reached the regular piece size.)
     if data:
-        pieces[i * 20:(i + 1) * 20] = sha1_20(data)
+        pieces[i * 20:(i + 1) * 20] = sha1(data)
 
     # Cut off unused pieces space.
     pieces = bytes(pieces[:(i + 1) * 20])
@@ -514,19 +556,6 @@ def get_best_trackers(count: int, url: str):
 
 
 def main() -> None:
-    # Validate the configuration.
-    for abbr, replacement in TRACKER_ABBR.items():
-        if not isinstance(abbr, str):
-            print("Configuration error: invalid tracker abbreviation: '%s' "
-                  "(must be a string instead)" % abbr,
-                  file=sys.stderr)
-            sys.exit(1)
-        if not isinstance(replacement, (str, list)):
-            print("Configuration error: invalid tracker abbreviation: '%s' "
-                  "(must be a string or list of strings instead)" % str(replacement),
-                  file=sys.stderr)
-            sys.exit(1)
-
     # Create and configure ArgumentParser.
     parser = argparse.ArgumentParser(
         description="py3createtorrent is a comprehensive command line utility for creating torrents.")
@@ -629,6 +658,11 @@ def main() -> None:
                         default=False,
                         help="include MD5 hashes in torrent file")
 
+    parser.add_argument("--config",
+                        type=str,
+                        action="store",
+                        help="use another config file instead of the default one from the home directory")
+
     parser.add_argument("-t",
                         "--tracker",
                         metavar="TRACKER_URL",
@@ -653,6 +687,29 @@ def main() -> None:
 
     args = parser.parse_args()
 
+    global VERBOSE
+    VERBOSE = args.verbose
+
+    config = Config()
+    if args.config:
+        if not os.path.isfile(args.config):
+            parser.error("The config file at '%s' does not exist" % args.config)
+        config.path = args.config
+
+    try:
+        config.load_config()
+    except json.JSONDecodeError as exc:
+        print("Could not parse config file at '%s'" % config.get_path_to_config_file(), file=sys.stderr)
+        print(exc, file=sys.stderr)
+        sys.exit(1)
+    except Config.InvalidConfigError as exc:
+        print(exc, file=sys.stderr)
+        sys.exit(1)
+
+    printv('Config / Tracker abbreviations:\n' + pprint.pformat(config.tracker_abbreviations))
+    printv('Config / Advertise:         ' + str(config.advertise))
+    printv('Config / Best trackers URL: ' + config.best_trackers_url)
+
     # Ask the user if he really wants to use uncommon piece lengths.
     # (Unless the force option has been set.)
     if not args.force and 0 < args.piece_length < 16:
@@ -676,9 +733,6 @@ def main() -> None:
     if args.verbose and args.quiet:
         parser.error("Being verbose and quiet exclude each other.")
 
-    global VERBOSE
-    VERBOSE = args.verbose
-
     # ##########################################
     # CALCULATE/SET THE FOLLOWING METAINFO DATA:
     # - info
@@ -695,7 +749,7 @@ def main() -> None:
         parser.error("'%s' neither is a file nor a directory." % input_path)
 
     # Evaluate / apply the tracker abbreviations.
-    trackers = replace_in_list(trackers, TRACKER_ABBR)
+    trackers = replace_in_list(trackers, config.tracker_abbreviations)
 
     # Remove duplicate trackers.
     trackers = remove_duplicates(trackers)
@@ -724,9 +778,10 @@ def main() -> None:
             m = regexp_best.match(t)
             if m:
                 try:
-                    new_trackers.extend(get_best_trackers(int(m.group(1)), BEST_TRACKERS_URL))
+                    new_trackers.extend(get_best_trackers(int(m.group(1)), config.best_trackers_url))
                 except urllib.error.URLError as e:
-                    print("Error: Could not download best trackers from '%s'. Reason: %s" % (BEST_TRACKERS_URL, e),
+                    print("Error: Could not download best trackers from '%s'. Reason: %s" %
+                          (config.best_trackers_url, e),
                           file=sys.stderr)
                     sys.exit(1)
             else:
@@ -818,7 +873,7 @@ def main() -> None:
     # - nodes (if at least one DHT bootstrap node was specified)
     # - creation date (may be disabled as well)
     # - created by
-    # - comment (may be disabled as well (if ADVERTISE = False))
+    # - comment (may be disabled as well)
 
     # Finish sub-dict "info".
     info['piece length'] = piece_length
@@ -874,7 +929,7 @@ def main() -> None:
     if isinstance(args.comment, str):
         if len(args.comment) > 0:
             metainfo['comment'] = args.comment
-    elif ADVERTISE:
+    elif config.advertise:
         metainfo['comment'] = "created with " + metainfo['created by']
 
     # Add the name field.
@@ -882,10 +937,10 @@ def main() -> None:
     if args.name:
         args.name = args.name.strip()
 
-        regexp = re.compile(r"^[A-Z0-9_\-., ]+$", re.I)
+        regexp = re.compile(r"^[A-Z0-9_\-., ()]+$", re.I)
 
         if not regexp.match(args.name):
-            parser.error("Invalid name: '%s'. Allowed chars: A_Z, a-z, 0-9, any of {.,_-} plus spaces." % args.name)
+            parser.error("Invalid name: '%s'. Allowed chars: A_Z, a-z, 0-9, any of {.,_-()} plus spaces." % args.name)
 
         metainfo['info']['name'] = args.name